requirements.txt
click==8.1.3
numpy==1.24.1
perlin_noise==1.12
Pillow==9.4.0
scikit_learn==1.2.1
